Gracefully catch errors in lazy component evaluation.
